requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.0
requests==2.31.0
httpx==0.27.0
pyyaml==6.0.1
openai
python-dotenv